Fixed few compilation issues
git-svn-id: https://omxil.svn.sourceforge.net/svnroot/omxil/libomxil/trunk@709 d11a7513-620d-0410-9d76-e80dcb6a9808

+++ src/omxcore.c
@@ -295,7 +295,7 @@ OMX_ERRORTYPE OMX_SetupTunnel(
   }
   DEBUG(DEB_LEV_PARAMS, "First stage of tunneling acheived:\n");
   DEBUG(DEB_LEV_PARAMS, "       - supplier proposed = %i\n", tunnelSetup->eSupplier);
-  DEBUG(DEB_LEV_PARAMS, "       - flags             = %i\n", tunnelSetup->nTunnelFlags);
+  DEBUG(DEB_LEV_PARAMS, "       - flags             = %i\n", (int)tunnelSetup->nTunnelFlags);
 
   component = (OMX_COMPONENTTYPE*)hInput;
   if (hInput) {
